requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.0
google-generativeai==0.8.3
gitpython==3.1.44
requests==2.32.3
